feat: 리다이렉트를 처리하기 위한 RedirectResponseTemplate 클래스 작성
