+++ app/src/main/java/nextstep/jwp/framework/http/template/StringResponseTemplate.java
@@ -8,15 +8,8 @@ public class StringResponseTemplate extends AbstractResponseTemplate {
 
     private static final String PLAIN_CONTENT_TYPE = "text/plain;charset=utf-8";
 
-    private static final String EMPTY_BODY = "";
-
     public HttpResponse ok(String body) {
         return template(HttpStatus.OK, new HttpHeaders(), body);
-    }
-
-    public HttpResponse found(String location) {
-        final HttpHeaders httpHeaders = new HttpHeaders().addHeader(HttpHeaders.LOCATION, location);
-        return template(HttpStatus.FOUND, httpHeaders, EMPTY_BODY);
     }
 
     @Override
